requirements.txt
memory-profiler==0.58.0
fastparquet==2024.11.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
numpy==2.2.3
pandas==2.2.3
plotly==6.0.0
pyarrow==19.0.1
pyodbc==5.2.0
schedule==1.2.2
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
SQLAlchemy==2.0.38
streamlit==1.43.0
streamlit-antd-components==0.3.2
streamlit-extras==0.5.5
streamlit-option-menu==0.4.0
streamlit-shadcn-ui==0.1.18
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
db==0.1.1
db-sqlite3
streamlit-autorefresh==1.0.1
flask
google-generativeai
prophet
waitress
slackclient
html2text
embedchain==0.1.128
streamlit-chat==0.1.1
sentence-transformers==4.1.0
pandasai
ollama
numpy==1.24.4